Server communication factored out
